+++ include/mp++/detail/fwd_decl.hpp
@@ -28,6 +28,12 @@ class real;
 
 #endif
 
+#if defined(MPPP_WITH_QUADMATH)
+
+class real128;
+
+#endif
+
 } // namespace mppp
 
 #endif